backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1

prometheus-fastapi-instrumentator==7.0.0

opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-logging==0.47b0

langchain==0.2.16
langgraph==0.2.40
faiss-cpu==1.8.0.post1
ragas==0.1.21

requests==2.32.3
fastapi
uvicorn[standard]
openai>=1.6.0
faiss-cpu
numpy
pydantic
python-dotenv
prometheus-fastapi-instrumentator
python-docx
streamlit
requests
pypdf
python-docx
numpy
python-dotenv
prometheus-fastapi-instrumentator==7.0.0
Prometheus-client
